feat: remove template model
