+++ backend/src/main/java/com/readme/rss/data/service/Impl/FrameworkServiceImpl.java
@@ -1,3 +1,6 @@
+// Client, Controller, Service : DTO로 전달
+// Service, DAO(Repository), DB : Entity로 전달
+
 package com.readme.rss.data.service.Impl;
 
 import com.readme.rss.data.dto.FrameworkDTO;
