+++ frontend/package.json
@@ -6,7 +6,6 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@uiw/react-markdown-editor": "^5.11.1",
     "@uiw/react-markdown-preview": "^4.1.10",
     "axios": "^1.3.4",
     "bootstrap": "^5.2.3",
@@ -19,7 +18,11 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
+    "rehype-stringify": "^9.0.3",
+    "remark-parse": "^10.0.1",
+    "remark-rehype": "^10.1.0",
     "styled-components": "^5.3.8",
+    "unified": "^10.1.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
